Add component for chatrooms
