+++ src/index.js
@@ -10,6 +10,7 @@ import { Home } from './Components/Home/home';
 import { Header} from './Components/header';
 import { Footer } from './Components/Footer/footer';
 import { ForgotPassword } from './Components/ForgotPassword/ForgotPassword';
+import { Chat } from './Components/Chatroom/chat';
 import { Chatroom } from './Components/Chatroom/chatroom';
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
@@ -21,7 +22,8 @@ root.render(
         <Route path="register" element={<SignUp />} />
         <Route path="dashboard" element={<Home />} />
         <Route path="forgot-password" element={<ForgotPassword/>} />
-        <Route path="/chatroom" element={<Chatroom />} />
+        <Route path="/chat" element={<Chat />} />
+        <Route path="/chatroom" element={<Chatroom />}/>
       </Routes>
     </BrowserRouter>
     <Footer />
