+++ src/utils/cable.js
@@ -1,6 +1,7 @@
 import { createConsumer } from "@rails/actioncable";
 
 const CableApp = {};
-CableApp.cable = createConsumer("ws://localhost:3001/cable");
+const token = localStorage.getItem("jwt");
+CableApp.cable = createConsumer(`ws://localhost:3001/cable?token=${token}`);
 
 export default CableApp;
